"To-Dos with Dates": mark challenge tasks as being from a challenge; turn dashboard tile yellow if the only "due" todos are old ones from challenges
From Tavern chat:
"@Alys Love the user display tool! I was surprised today, when going excitedly to view the new to-dos feature, that it said I had a to-do due. Turns out the due date was a while back, but I can't delete it because it's from a still-ongoing challenge. Not sure if that's something that could be changed or if it's even useful to change it (it doesn't bother me now that I know what caused it). Besides, the list of to-do dates in chronological order is the really useful part.  Incredibly useful. I might even start adding things with really firm due dates to my to-do list on Habit instead of just Todoist... -- Clare-Dragonfly"

The ideal situation would be to hide any old challenge to-do that you don't care about but emphasise (as now) any that you do care about, but of course it's impossible for the tool to know which is which. :) What I might do is keep it working exactly as it does now, except that the dashboard tile will turn yellow instead of red if the ONLY "due" todos are from challenges AND are more than a week old (i.e., PROBABLY challenge todos that you don't care about). That way you would be able to ignore the dashboard tile when it's yellow, but anyone who does still care about the old challenge todo will still have that visual reminder of them. How does that sound? -- Alys"

"@Alys That sounds awesome! You are awesome. -- Clare-Dragonfly"