Use custom legend and limit amount of data in chart
The custom legend solves the emoji problem since I
elect to not escape the html special chars. It also
frees up chart space for the actual chart.

I'm kind of bummed I had to limit the data in the chart.
The way it is now, we could actually just drop all of the
data in there and it would look fine, but the page would
be extremely janky. Let me rephrase - the user's computer
would be extremely janky (just watch your CPUs scream).