Use a timeseries instead of categories for accurate scale
Since I had the data organized as categorical the timescale
along the x-axis was misleading. I pivoted the data so that
I could have a proper timeseries. Many thanks to jtleek for
this guide which lead me to tidy up the data:

https://github.com/jtleek/datasharing#the-tidy-data-set

I also changed the default value for when data is missing
for a time entry. Prior to this I had just defaulted to 0,
so you would see a spike in the graph each time someone
clicked a habit. Now, I use the value of the previous entry.
This way you can clearly see the trend of the habit over time.
It really serves as a poor-man's interpolation for now...