Remove double-sanitization which foiled markdown lexing
The marked library does it own santization. What we had
already was escaping entities that are necessary for proper
lexing of markdown. In the instance I found it, the entity
that I needed was the double quote. Since it had been escaped,
it threw off the translation of an image with title text.

But don't worry:

    -> marked("&<>\"'", {sanitize: true})
    <- "<p>&amp;&lt;&gt;&quot;&#39;</p>
       "