Render markdown in task notes as well
This takes care of sanitizing and rendering
the markdown that people put in notes. This
is particularly useful, since even I throw
links in the notes and expect them to render.

AFAIK, that's the last of the fields we need
to render/santize for markdown. At least I
don't remember other fields I type in on
HabitRPG that show up here? I thought maybe
the user profile, but I don't see that on the
page.