Add animations to dashboard and move value-only items
Having the value-only items intermingle with the charts
looked awkward, so I separated them. The DASHBOARD now
has 2 list items, visualized and valueOnly, which then
contain lists of their own.

I also changed the gauges and pie to initialize with
no data, then have a setTimeout load the data later.
This results in a nice animation showing the gauges fill.
The best part is the pie chart showing the damage occuring,
which I think makes the meaning abundantly clear.