temporarily revert wikia link code to the external format
This is so I can use my current output testing process while I
improve how the wikia and html output is produced, so that we can
hopefully do away with the regexes for converting link types. Once
that's done I'll change the wikia links back to internal format.

This also updates the version date.