refactor code for creating `rowHtml`
It's no longer necessary to use regexes for replacing wiki links
with html links for the HTML table, but some other ugly hacks had
to be added.
More refactoring is needed.

The basic problem is that the HTML table was added as an afterthought
when I first made this page and it's really showing now.