Bug fix: Applied user's timezone to dailies history.
This fixes a bug where the dailies history was showing incorrect dates for some users. 

# Changes:
- The user's current time zone offset is now retrieved from the user preferences.
- A local date (in the users time zone), is calculated from the history date.
Note: The local date is only "local" to a human reader if displayed in the UTC zone. This is a workaround to limitations of the built in JavaScript date handling functions.
- The local date is formatted and displayed, using a new UTC date formatting function. 'formatDateFromUTC()'
(The new formatting function was adapted from the old 'myDateConverter()'.)
- The new formatting function creates a human readable string from UTC time zone values.  

# About the date handling in javascript:
Apparently JavaScript does not make it easy to format dates or timestamps to a specified time zone offset. 

For example...
Date.getHours() Returns the hours in the local time of the "JavaScript environment".
Date.getUTCHours() Returns the hours in the UTC time zone.
Date.toLocaleString('en-US', { timeZone: 'America/New_York' })
This one can format a date with a specified IANA time zone identifier. (But this function can not take an offset value.)

None of the above functions does what was wanted here. We wanted to format a timestamp as a human readable string with a specified offset value.

The implemented solution was adapted from Denis Howe's answer at:
https://stackoverflow.com/questions/12413243/javascript-date-format-like-iso-but-local